refactor(architecture): close 3 of 4 ADR-061 Phase 5 item 2 facade exceptions
poi.py's is_public_dependency_node dependency is physically relocated to
model/source_graph_query.py (pure, model-only logic, same shape
model/mangled_name.py already established), unblocking the extract ->
compare edge check_architecture.py forbids. cli_buildsource_helpers.py/
cli_buildsource_merge.py reach build_source_graph/
mark_source_edges_extractor_coverage through workflows/extraction.py's
re-export block, the same sanctioned frontends -> workflows -> extract
indirection pack_io.py already uses. template_graph.py remains, blocked
by an unrelated 2000-line hard-cap constraint.
