+++ changelog.d/20260831_195417_noreply_worktree_agent_a22558915b79b93da.md
@@ -1,0 +1,29 @@
+<!--
+A new changelog fragment. See changelog.d/README.md for the workflow.
+
+Uncomment exactly ONE '### <Category>' section below (remove its comment
+wrapper) and replace the example bullet with your entry, written the way
+it should read in CHANGELOG.md. Delete the other sections.
+-->
+
+### Changed
+
+- **Closed three of the four ADR-061 Phase 5 item 2 facade exceptions** —
+  `buildsource/poi.py`, `cli_buildsource_helpers.py`, and
+  `cli_buildsource_merge.py` no longer need `buildsource/source_graph.py`'s
+  temporary compatibility facade. The shared graph node/edge-classification
+  predicates (`is_public_dependency_node` and its siblings) physically moved
+  from the `compare`-classified `buildsource/source_graph_query.py` into
+  `abicheck/model/source_graph_query.py` — pure, dependency-free logic over
+  `model`'s own data, the same shape `model/mangled_name.py` already
+  established for `itanium_scope_components` — so `extract`-classified
+  `poi.py` can import it directly (every layer may import `model`).
+  `build_source_graph`/`mark_source_edges_extractor_coverage` (`extract`)
+  are now re-exported from `workflows/extraction.py`, the sanctioned
+  `frontends -> workflows -> extract` indirection already used for
+  `pack_io.py`, so the two `frontends`-classified `cli_buildsource_*.py`
+  modules reach them without a forbidden `frontends -> extract` edge.
+  `buildsource/source_graph_query.py` is now a thin back-compat re-export
+  facade. No behavior change; `python scripts/check_architecture.py` stays
+  at 0 errors.
+
